refactor(Lotto) : 로또 테스트 파일 합치기
